Cross-platform path handling
Removes the `zarr_path` variable from BaseZarrLocation, using a
pathlib.Path field in LocalZarrLocation and a URL string in
RemoteZarrLocation to properly handle the creation of subpaths across
multiple platforms. In general, use of os.path and especially
os.path.join should be avoided in favor of delegating to methods on the
zarr location objects.