Deactivate Zarr dev build
This build was put in place by the Zarr team in order to check
for potential issues in downstream projects after PRs have been
merged. For some reason, it is now aborting with little to no
logging. In https://github.com/zarr-developers/zarr-python/issues/1328
the `texasbbq` package from Numba has been suggested for a similar
purpose. This workflow might could be rewritten using such a tool.